Basic authentication/authorization mechanism added using Spring Security (for learning purposes)
Just added security-starter to classpath in pom.xml and that gives me
basic authentication magically. Then configured a
WebSecurityConfigurerAdapter at WebSecurityConfig.java file that added
me the power creating a basic authentication (user, password and role)
and authorization (what role has access to what resource -- in this case
a path) scheme.

In frontend I finally started to use ngResource to call the rest
services in the backend, instead of using $http calls. Created the
ATMServices.js to store the rest calling services.

Unused classes in the backend:
- UserManagementController (was supposed to be the one who would
authenticate, but could not make it work so far. Tried to use token
based authentication/authorization but without any success).

Unused code in frontend:
- ATMInterceptors.js: was supposed to intercept the call and include the
token in the header, but at the end of the day, I could not use token
based authentication, so the interceptor (which is working) was rendered
useless.